rename directory transfer -> transfers for consistency
