+++ CMakeLists.txt
@@ -73,11 +73,11 @@ SET(TARGET_SRC
      include/exadg/solvers_and_preconditioners/solvers/enum_types.cpp
      include/exadg/solvers_and_preconditioners/multigrid/multigrid_preconditioner_base.cpp
      include/exadg/solvers_and_preconditioners/multigrid/multigrid_input_parameters.cpp
-     include/exadg/solvers_and_preconditioners/multigrid/transfer/mg_transfer_p.cpp
-     include/exadg/solvers_and_preconditioners/multigrid/transfer/mg_transfer_c.cpp
-     include/exadg/solvers_and_preconditioners/multigrid/transfer/mg_transfer_h.cpp
-     include/exadg/solvers_and_preconditioners/multigrid/transfer/mg_transfer_global_coarsening.cpp
-     include/exadg/solvers_and_preconditioners/multigrid/transfer/mg_transfer_global_refinement.cpp
+     include/exadg/solvers_and_preconditioners/multigrid/transfers/mg_transfer_p.cpp
+     include/exadg/solvers_and_preconditioners/multigrid/transfers/mg_transfer_c.cpp
+     include/exadg/solvers_and_preconditioners/multigrid/transfers/mg_transfer_h.cpp
+     include/exadg/solvers_and_preconditioners/multigrid/transfers/mg_transfer_global_coarsening.cpp
+     include/exadg/solvers_and_preconditioners/multigrid/transfers/mg_transfer_global_refinement.cpp
      include/exadg/postprocessor/error_calculation.cpp
      include/exadg/postprocessor/mean_scalar_calculation.cpp
      include/exadg/postprocessor/lift_and_drag_calculation.cpp
