+++ include/exadg/solvers_and_preconditioners/multigrid/multigrid_preconditioner_base.cpp
@@ -36,8 +36,8 @@
 #include <exadg/solvers_and_preconditioners/multigrid/smoothers/chebyshev_smoother.h>
 #include <exadg/solvers_and_preconditioners/multigrid/smoothers/gmres_smoother.h>
 #include <exadg/solvers_and_preconditioners/multigrid/smoothers/jacobi_smoother.h>
-#include <exadg/solvers_and_preconditioners/multigrid/transfer/mg_transfer_global_coarsening.h>
-#include <exadg/solvers_and_preconditioners/multigrid/transfer/mg_transfer_global_refinement.h>
+#include <exadg/solvers_and_preconditioners/multigrid/transfers/mg_transfer_global_coarsening.h>
+#include <exadg/solvers_and_preconditioners/multigrid/transfers/mg_transfer_global_refinement.h>
 #include <exadg/solvers_and_preconditioners/utilities/compute_eigenvalues.h>
 
 namespace ExaDG
