+++ include/exadg/solvers_and_preconditioners/multigrid/transfers/mg_transfer_global_refinement.h
@@ -29,7 +29,7 @@
 #include <deal.II/multigrid/mg_constrained_dofs.h>
 
 // ExaDG
-#include <exadg/solvers_and_preconditioners/multigrid/transfer/mg_transfer.h>
+#include <exadg/solvers_and_preconditioners/multigrid/transfers/mg_transfer.h>
 
 namespace ExaDG
 {
